Review basebackup script

+++ files/scripts/basebackup.sh
@@ -4,19 +4,6 @@ set -e
 
 test -d "${PGDATA}"
 
-if [ -n "$(ls -A ${PGDATA})" ]; then
-    echo "The target directory (${PGDATA}) is not empty. Nothing to do." 1>&2
-    if ! [ -f "${PGDATA}/PG_VERSION" ]; then
-        echo "The target directory (${PGDATA}) does not seem like a PostgreSQL data directory!" 1>&2
-        exit 1
-    fi
-    if ! [ -f "${PGDATA}/recovery.conf" ]; then
-        echo "The data directory does not contain recovery.conf!" 1>&2
-        exit 1
-    fi
-    exit 0
-fi
-
 # Read command line
 
 archive_dir=/var/backups/postgresql/archive
@@ -53,22 +40,42 @@ replication_password="$(cat ${replication_password_file})"
 
 pgpass_file=~/.pgpass
 touch ${pgpass_file}
-chmod 0600 ${pgpass_file}
+chmod u=rw,g=,o= ${pgpass_file}
 
 pgpass_line=$(echo -n "${master_host}:5432:*:${replication_user}:${replication_password}")
 grep -qF -e "${pgpass_line}" ${pgpass_file} || echo "${pgpass_line}" >> ${pgpass_file}
 
 # Take basebackup
 
-pg_basebackup -v --checkpoint=fast -h ${master_host} -U ${replication_user} -D ${PGDATA}
+if [ -n "$(ls -A ${PGDATA})" ]; then
+    echo "The target directory (${PGDATA}) is not empty. Skipping basebackup." 1>&2
+    if ! [ -f "${PGDATA}/PG_VERSION" ]; then
+        echo "The target directory (${PGDATA}) is not a data directory! (PG_VERSION not found)" 1>&2
+        exit 1
+    fi
+    if [ -f "${PGDATA}/recovery.done" ]; then
+        echo "The data directory has finished recovery!" 1>&2
+        exit 1
+    fi
+else
+    # data directory is empty
+    echo "Taking a basebackup from ${master_host}..."  1>&2
+    pg_basebackup -v --checkpoint=fast -h ${master_host} -U ${replication_user} -D ${PGDATA}
+fi
 
 # Generate recovery.conf
 
-cat <<EOD > "${PGDATA}/recovery.conf"
-standby_mode = 'on'
-primary_conninfo = 'user=${replication_user} password=''${replication_password}'' host=${master_host} port=5432 sslmode=prefer sslcompression=0 krbsrvname=postgres target_session_attrs=any'
-trigger_file = 'trigger-failover'
-restore_command = 'test ! -f ${archive_dir}/%f || cp -v ${archive_dir}/%f %p'
-EOD
+if ! [ -f "${PGDATA}/recovery.conf" ]; then
+    echo "Generating recovery.conf..." 1>&2
+    
+    cat <<-EOD > "${PGDATA}/recovery.conf"
+	standby_mode='on'
+	primary_conninfo='user=${replication_user} password=''${replication_password}'' host=${master_host} port=5432 sslmode=prefer sslcompression=0'
+	trigger_file='trigger-failover'
+	restore_command='test ! -f ${archive_dir}/%f || cp -v ${archive_dir}/%f %p'
+	EOD
 
-chmod 0660 "${PGDATA}/recovery.conf"
+    chmod u=rw,g=rw,o= "${PGDATA}/recovery.conf"
+fi
+
+
